+++ packages/components-js/CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Description property to `p-text-field-wrapper`, `p-textarea-wrapper` and `p-select-wrapper` 
 
+### Changed
+- Improve accessibility of error and success states of form elements
+- Aria-invalid attribute of form elements if they are in error state is now managed by component
+
 ## [1.4.0-rc.0] - 2020-04-09
 
 ### Added